requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
pathspec==0.8.0
pylint==2.5.3
pyparsing==2.4.7
pytz==2020.1
regex==2020.7.14
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
